+++ src/main/java/org/ppijerman/ppijidentitybackend/server/service/security/dao/IpRateLimitedException.java
@@ -1,4 +1,4 @@
-package org.ppijerman.ppijidentitybackend.server.service.security;
+package org.ppijerman.ppijidentitybackend.server.service.security.dao;
 
 import org.springframework.security.core.AuthenticationException;
