+++ src/main/java/org/ppijerman/ppijidentitybackend/server/service/security/dao/PpijIdAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package org.ppijerman.ppijidentitybackend.server.service.security;
+package org.ppijerman.ppijidentitybackend.server.service.security.dao;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
